fr: mc-mitm + : before/after stop separation t: mc_mitm change channel parsing

+++ wpa3_test/include/attacks/mc_mitm/client_state.h
@@ -8,8 +8,8 @@ namespace wpa3_tester{
 
     struct NetworkConfig {
         std::string ssid;
-        int real_channel  = -1;
-        int rogue_channel = -1;
+        uint8_t real_channel  = -1;
+        uint8_t rogue_channel = -1;
     };
 
     class ClientState {
